moves sharded locks to common package
