+++ adapters/repos/db/vector/cache/sharded_lock_cache.go
@@ -19,13 +19,12 @@ import (
 	"unsafe"
 
 	"github.com/sirupsen/logrus"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector"
 	"github.com/weaviate/weaviate/adapters/repos/db/vector/common"
 	"github.com/weaviate/weaviate/adapters/repos/db/vector/hnsw/distancer"
 )
 
 type shardedLockCache[T float32 | byte | uint64] struct {
-	shardedLocks     *vector.ShardedLocks
+	shardedLocks     *common.ShardedLocks
 	cache            [][]T
 	vectorForID      common.VectorForID[T]
 	normalizeOnRead  bool
@@ -66,7 +65,7 @@ func NewShardedFloat32LockCache(vecForID common.VectorForID[float32], maxSize in
 		maxSize:          int64(maxSize),
 		cancel:           make(chan bool),
 		logger:           logger,
-		shardedLocks:     vector.NewDefaultShardedLocks(),
+		shardedLocks:     common.NewDefaultShardedLocks(),
 		maintenanceLock:  sync.RWMutex{},
 		deletionInterval: deletionInterval,
 	}
@@ -86,7 +85,7 @@ func NewShardedByteLockCache(vecForID common.VectorForID[byte], maxSize int,
 		maxSize:          int64(maxSize),
 		cancel:           make(chan bool),
 		logger:           logger,
-		shardedLocks:     vector.NewDefaultShardedLocks(),
+		shardedLocks:     common.NewDefaultShardedLocks(),
 		maintenanceLock:  sync.RWMutex{},
 		deletionInterval: deletionInterval,
 	}
@@ -106,7 +105,7 @@ func NewShardedUInt64LockCache(vecForID common.VectorForID[uint64], maxSize int,
 		maxSize:          int64(maxSize),
 		cancel:           make(chan bool),
 		logger:           logger,
-		shardedLocks:     vector.NewDefaultShardedLocks(),
+		shardedLocks:     common.NewDefaultShardedLocks(),
 		maintenanceLock:  sync.RWMutex{},
 		deletionInterval: deletionInterval,
 	}
