+++ adapters/repos/db/vector/common/sharded_locks_test.go
@@ -9,7 +9,7 @@
 //  CONTACT: hello@weaviate.io
 //
 
-package vector
+package common
 
 import (
 	"sync"
