+++ adapters/repos/db/vector/hnsw/vector_cache_prefiller_test.go
@@ -17,7 +17,7 @@ import (
 
 	"github.com/sirupsen/logrus/hooks/test"
 	"github.com/stretchr/testify/assert"
-	"github.com/weaviate/weaviate/adapters/repos/db/vector"
+	"github.com/weaviate/weaviate/adapters/repos/db/vector/common"
 )
 
 func TestVectorCachePrefilling(t *testing.T) {
@@ -25,7 +25,7 @@ func TestVectorCachePrefilling(t *testing.T) {
 	index := &hnsw{
 		nodes:               generateDummyVertices(100),
 		currentMaximumLayer: 3,
-		shardedNodeLocks:    vector.NewDefaultShardedLocks(),
+		shardedNodeLocks:    common.NewDefaultShardedLocks(),
 	}
 
 	logger, _ := test.NewNullLogger()
